✨ feat: 한  줄 메모 명함 리스트 - 선택 디자인 구현 완료
